Move IDEUtil class to miglayout-core and delete obsolete ideutil project
The IDEUtil class is used to expose package-private functionality from
miglayout-core to external projects.
This mechanism fails in an OSGi environment, because the IDEUtil class
is loaded by a different classloader that is used to load classes it
tries to access.

To solve this problem, move IDEUtil to miglayout-core. Given that the
ideutil project is now empty, it is removed from the repository.

resolves #103